refactor: setErrorMessag의 위치를 utils에서 hook으로 변경
